refactor: improves folder naming
