+++ __tests__/__mock_helpers__/fake_hathorwallet.ts
