+++ __tests__/pushNotification.test.ts
@@ -1,5 +1,5 @@
-import { mockAxiosAdapter } from './__mock_helpers/axios-adapter.mock';
-import { buildWalletToAuthenticateApiCall } from './__mock_helpers/wallet-service.fixtures';
+import { mockAxiosAdapter } from './__mock_helpers__/axios-adapter.mock';
+import { buildWalletToAuthenticateApiCall } from './__mock_helpers__/wallet-service.fixtures';
 import { PushNotification, PushNotificationProvider } from '../src/pushNotification';
 import config from '../src/config';
